Reconocimiento-de-caracteres-braile-CNN/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
numpy
python-multipart
ipykernel
opencv-python
matplotlib
scipy